@@ -1,23 +1,18 @@
 <template>
   <div v-on:mouseover="isActive = true" v-on:mouseleave="isActive = false" class="flex space-x-2 flex-grow cursor-pointer">
     <slot :isActive="isActive"></slot>
-    <button @click="$emit('clicked')" class="text-white desktop:text-lg laptop:text-md">{{ title }}</button>
+    <button @click="$emit('clicked')" class="text-white desktop:text-lg laptop:text-md">{{ props.title }}</button>
   </div>
 </template>
 
-<script>
-export default {
-  name: "SidebarButton",
-  emits: ['clicked'],
-  data() {
-    return {
-      isActive: false
-    }
-  },
-  props: {
+<script setup>
+  import { defineEmits, ref, defineProps } from 'vue'
+
+  defineEmits(['clicked'])
+  const isActive = ref(false)
+  const props = defineProps({
     title: String,
     image: String
-  }
-}
+  })
 </script>
 
