@@ -1,9 +1,6 @@
-import Vue from 'vue'
+import {createApp} from 'vue'
 import App from './App.vue'
-import './assets/tailwind.css'
+import './main.css'
 
-Vue.config.productionTip = false
 
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+createApp(App).mount('#app')
