@@ -1,22 +1,19 @@
 <template>
   <div class="w-full">
-    <label class="block mb-2 laptop:text-sm desktop:text-md" for="text input">{{ title }}</label>
+    <label class="block mb-2 laptop:text-sm desktop:text-md" for="text input">{{ props.title }}</label>
     <input class="outline-none apperance-none border-2 rounded-full bg-secondary border-border py-2 px-4" :class="widthFormat" alt="Text Input">
   </div>
 </template>
 
-<script>
-export default {
-  props: {
+<script setup>
+  import { computed, defineProps } from 'vue'
+
+  const props = defineProps({
     title: String,
     width: String
-  },
-  computed: {
-    widthFormat() {
-      return `w-${this.width}`
-    }
-  }
-}
+  })
+  const widthFormat = computed(() => `w-${props.width}`)
+
 </script>
 
 
